Wrap internal implementation of PasswordRAMStore in the unnamed namespace
PasswordRAMStore class was already coded using private implementation
PWStore class.  Wrap PWStore class in the unnamed namespace so the
compiler makes the type local to the translation unit [1][2].

The single_pwstore object was already local because of it's static
qualifier.  It is included in the unnamed namespace for completeness but
this makes no difference because the static qualifier and being in the
unnamed namespace are both making the variable local to the translation
unit.

Keep the static qualifier as a per variable / per function reminder of
local linkage.

[1] Why should you prefer unnamed namespaces over static functions?
    https://stackoverflow.com/questions/154469/why-should-you-prefer-unnamed-namespaces-over-static-functions
        "Static only applies to names of objects, functions, and
        anonymous unions, not to type declarations.
        "

[2] How are unnamed namespaces superior to the static keyword?
    [duplicate]
    https://stackoverflow.com/questions/4422507/how-are-unnamed-namespaces-superior-to-the-static-keyword
        "unnamed namespace's are superior to the static keyword,
        primarily because the keyword static applies only to the
        variables declarations and functions, not to the user-defined
        types.
        "