Convert DMRaid class to a static interface (!135)
The DMRaid module had a pseudo multi-object interface and used the
constructor to load the cache.  However all the data in the class is
static so a DMRaid object is empty.  But a DMRaid object was still
needed just to call the member functions.

Make all the member functions static and provide an explicit
load_cache() method.

Previous examples of doing this:
[1] commit a94be2da05c65b36c290d993c89f5580e85042ca
    Switch to static class interface for FS_Info
[2] commit 912766c2e11aab2cecd916bdc2027cc14087c4a0
    Switch to a static interface for Proc_Partitions_Info

Closes !135 - Refactor DMRaid to a static class interface and make
              mutating getters return by value