Avoid theoretical nullptr dereference in File System Support dialog (#282)
Now that the previous commit accepts that looking up a stock icon may
fail, which results in Utils::mk_pixbuf() returning the nullptr, the
rest of the GParted code needs to handle this possibility too.  With
this debugging change applied, mk_pixbuf() always returns a nullptr:
    --- a/src/Utils.cc
    +++ b/src/Utils.cc
@@ -176,7 +176,8 @@ Glib::RefPtr<Gdk::Pixbuf> Utils::mk_pixbuf(Gtk::Widget& widget,
                                                const Gtk::StockID& stock_id,
                                                Gtk::IconSize icon_size)
     {
    -    Glib::RefPtr<Gdk::Pixbuf> theme_icon = widget.render_icon_pixbuf(stock_id, icon_size);
    +    static const Gtk::BuiltinStockID UNKNOWN = { ((GtkStock)"unknown") };
    +    Glib::RefPtr<Gdk::Pixbuf> theme_icon = widget.render_icon_pixbuf(UNKNOWN, icon_size);
         if (! theme_icon)
             return theme_icon;

In the DialogFeatures constructor icon_yes and icon_no are now assigned
the nullptr which it then dereferences with icon_yes->get_width() and
icon_yes->get_height(), causing a crash.  Fix the code to be safe
against this possibility.

Closes #282 - Crash due to not checking for failure to load icon