Define and initialise local variables one per line
Happened to be grepping the code for ' = 0' (assign or initialise to 0)
and these cases took extra investigation because they didn't follow the
pattern for local variables of define and initialise one variable per
statement and per line.  Make them fit that pattern.

Many other matches were found by the grep but they mostly seemed to be
subsequent assignment of local or member variables after initialisation
so not applicable to following this pattern.