Rename member variable to DrawingAreaVisualDisk::m_visual_partitions
More work towards having member variables start with m_ to make it
abundantly clear that they are members of a class.

Also for member variable visual_partitions, set_static_data() and
set_selected() methods receive a parameter of the same name which
shadows it.  This is confusing and bad practice [1].  set_static_data()
and set_selected() are recursive to handle logical partitions so must
receive the vector as a parameter, so can't be changed to directly use
the member variable.  The parameter could be renamed but what would be a
good name for the vector of visual partitions which is
m_visual_partitions except when it's a sub-part.  (Rhetorical question).

[1] C++ Core Guidelines, ES.12: Do not reuse names in nested scopes
    https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#es12-do-not-reuse-names-in-nested-scopes
        Reason: It is easy to get confused about which variable is used.
        Can cause maintenance problems.
        ...
        Enforcement:
        ...
        * Flag reuse of a member name as a local variable in a member
          function