Separate FSCK type messages from the label reported by fatlabel (#286)
Following a lockup and reboot of my desktop while several USB keys were
mounted GParted is now reporting fsck.fat messages as part of the
FAT16/32 file system label for those partitions.

For the FAT32 file system on the first USB key:
(Run fsck.fat and fatlabel throwing away stderr to demonstrate the
messages aren't being written to that stream):
    # fsck.fat -n /dev/sda1 2> /dev/null
    fsck.fat 4.1 (2017-01-24)
    There are differences between boot sector and its backup.
    This is mostly harmless. Differences: (offset:original/backup)
      71:42/4e, 72:52/4f, 73:49/20, 74:44/4e, 75:47/41, 76:45/4d, 77:54/45
      , 78:53/20, 79:5f/20, 80:4b/20, 81:31/20
      Not automatically fixing this.
    /dev/sda1: 82 files, 1489643/1956005 clusters

    # fatlabel /dev/sda1 2> /dev/null
    There are differences between boot sector and its backup.
    This is mostly harmless. Differences: (offset:original/backup)
      71:42/4e, 72:52/4f, 73:49/20, 74:44/4e, 75:47/41, 76:45/4d, 77:54/45
      , 78:53/20, 79:5f/20, 80:4b/20, 81:31/20
      Not automatically fixing this.
    BRIDGETS_K1

For the FAT32 file system on the second USB key:
    # fsck.fat -n /dev/sdb1 2> /dev/null
    fsck.fat 4.1 (2017-01-24)
    0x41: Dirty bit is set. Fs was not properly unmounted and some data may be corrupt.
     Automatically removing dirty bit.
    Leaving filesystem unchanged.
    /dev/sdb1: 342 files, 101636/1890667 clusters

    # fatlabel /dev/sdb1 2> /dev/null
    0x41: Dirty bit is set. Fs was not properly unmounted and some data may be corrupt.
     Automatically removing dirty bit.
    MIKES-KEY2

Fix this by separating the fatlabel output using the heuristic that the
label is on the last line and anything before that is an FSCK type
message.  According to the FAT Specification [1] control characters,
including the new line character, are illegal in short directory
entries, including the volume label.  GParted, mlabel and fatlabel 4.2
and later don't allow putting new lines characters into labels, but
fatlabel 4.1 and earlier does.  It seems very unlikely that anyone will
have put new line characters into a FAT16/32 label and this will become
less possible as distributions still using dosfstools 4.1 reach end-of-
life.  This simple heuristic feels like the right solution.
    # fatlabel /dev/sda1 | tail -1
    BRIDGETS_K1

    # fatlabel /dev/sdb1 | tail -1
    MIKES-KEY2

Set the FSCK type messages as a partition message which GParted displays
to the user as a partition warning.  Running FSCK on these FAT32 file
systems fixes them, clearing the errors.  This can be done in GParted
using the check operation.

[1] Microsoft FAT Specification
    https://academy.cba.mit.edu/classes/networking_communications/SD/FAT.pdf
        6.1 FileDirectory Name
        ...
        Restrictions on characters comprising the name:
        * Lower case characters are not allowed
        * Illegal values for characters in the name are as follows:
          * Values less than 0x20 (except for the special case of 0x05
            in DIR_Name[0] described earlier)

Closes #286 - GParted 1.7.0-git displays error messages along with the
              FAT label