Rename treeview_flags_[Cc]olumns struct and member variable
    $ egrep 'treeview_[a-z]*_Columns treeview_[a-z]*_columns' include/*.h
    include/DialogManageFlags.h:  treeview_flags_Columns treeview_flags_columns ;
    include/Dialog_Progress.h:    treeview_operations_Columns treeview_operations_columns;
    include/HBoxOperations.h:     treeview_operations_Columns treeview_operations_columns;
    include/TreeView_Detail.h:    treeview_detail_Columns treeview_detail_columns;
    include/Win_GParted.h:        treeview_devices_Columns treeview_devices_columns ;

Having a struct and a member variable only differing by the case of a
single character towards the end of the name is very hard to spot.  To
better distinguish them rename them.  Use the GParted predominant
"PascalCase" / "Pascal_Snake_Case" [1] for the struct and "snake_case"
for the variable.  Additionally prefix the variable name with m_ to make
it clear it's a member.

[1] Naming convention (programming),
    Examples of multiple-word identifier formats
    https://en.wikipedia.org/wiki/Naming_convention_(programming)#Examples_of_multiple-word_identifier_formats