Fix configure script detection of C++14 compiler (!141)
As part of merge request !138 "Require C++14 compilation and update to
Google Test 1.16.0" GParted was successfully built by the GitLab
Continuous Integration jobs on Alpine Linux 3.17.  But looking closely
at the job output showed this:
    $ ./autogen.sh
    ...
    Running ./configure --enable-maintainer-mode ...
    ...
    checking for glibmm-2.4 >= 2.32.0... yes
    yes
    checking for gtk_show_uri_on_window() function... yes

Where as it should have be:
    $ ./autogen.sh
    ...
    Running ./configure --enable-maintainer-mode ...
    ...
    checking for glibmm-2.4 >= 2.32.0... yes
    checking whether g++ supports C++14 features by default... yes
    checking for Gtk::ScrolledWindow::set_propagate_natural_width() method... yes
    checking for gtk_show_uri_on_window() function... yes

The generated configure script didn't check for a standard C++14
compiler and didn't print the "checking for
Gtk::ScrolledWindow::set_propagate_natural_width() method..." message.
This is because Alpine Linux doesn't have the autoconf-archive package
installed by default to provide the AX_CXX_COMPILE_STDCXX_14 macro and
GParted doesn't include a local copy.  So the call to the macro was
passed through from configure.ac to the configure shell script
unchanged.  The fragment of generated configure shell script looked like
this:
    AX_CXX_COMPILE_STDCXX_14()

    { printf "%s\n" "$as_me:${as_lineno-$LINENO}: checking for Gtk::ScrolledWindow::set_propagate_natural_width() method" >&5
    printf %s "checking for Gtk::ScrolledWindow::set_propagate_natural_width() method... " >&6; }

Instead of having shell code to check for the presence of a C++14
standard compiler the macro name became the start of a shell function
and the following two curly-brace enclosed lines, which were meant to
print the missing following message, became it's body.

This is only an issue on distributions which don't have autoconf-archive
installed and GParted is compiled from the git repository, using
./autogen.sh.  Fix this by adding a copy of the AX_CXX_COMPILE_STDCXX_14
macro and updating to the latest AX_CXX_COMPILE_STDCXX macro, which it
calls.

Closes !141 - Fix configure script detection of C++14 compiler when
              generated without AX_CXX_COMPILE_STDCXX_14 being available