Disallow operation merging past others on the same partition (#117)(#780034)
A user reported that after two partition name operations were merged
together past other intervening operations on the same partition, the
main window previewed the partition with the first, rather than the
second name.

For example compose these operations all on the same partition:
1. Set partition name "ONE" on /dev/sdb1
2. Set file system label "NEWLABEL" on /dev/sdb1
3. Set partition name "TWO" on /dev/sdb1

GParted merges the 3rd operation into the 1st giving this display in the
main window:
    Partition   | Name | File System | Label
    -----------------------------------------------
    /dev/sdb1     ONE           ext4   NEWLABEL
    unallocated          unallocated

    -----------------------------------------------
    * Set partition name "TWO" on /dev/sdb1
    * Set file system label "NEWLABEL" on /dev/sdb1

    -----------------------------------------------
    2 operations pending

When GParted composes each operation it saves a before and after copy of
the partition object [1].  When GParted merges an operation it updates
the earlier operation's after partition object with that from the
operation being merged from [2].  And when GParted visually applies the
pending operations it replaces the current displayed partition object
with the after copy recorded in the operation [3].

[1] Operation::partition_original and ::partition_new respectively
[2] Operation*::merge_operations()
[3] Operation*::apply_to_visual()

So after merging the two remaining operations visually applied this:
1. Replaced the partition object for display with one containing the
   partition name of "TWO", as updated by the operation merging.
2. Replaced the partition object for display with one containing the
   file system label of "NEWLABEL", but also the partition name of "ONE"
   because that is what it was when the 2nd operation was composed.

Therefore it is not correct to simply replace the partition object when
visually applying the operations after operations have been merged past
others on the same partition.  So far this is only a visual preview
issue.

Consider this new test case of composing the following operations
affecting the same partition:
1. Set file system label "ONE" on /dev/sdb1
2. Copy /dev/sdb2 to /dev/sdb1
3. Set file system label "TWO" on /dev/sdb1

GParted merges the 3rd operation into the 1st leaving these operations:
1. Set file system label "TWO" on /dev/sdb1
2. Copy /dev/sdb2 to /dev/sdb1

These two sequences of operations will not have the same affect when
applied.  Without merging the label of sdb1 would be "TWO" but with
merging the label of sdb1 would be whatever the label of the file system
copied from sdb2 was.

This specific test case was chosen because it can and has been tested on
the current GParted master and GParted 0.10.0 (when operation merging
was first introduced).  In general it is not correct to merge operations
past other operations affecting the same partition.

Partition naming is a property of the partition so could be merged past
other file system specific manipulation operations on the same
partition.  However file system operations such as new UUID and label
file system can't be merged past paste into the partition because those
all affect the partition contents.  Don't try to work out and code the
complicated rules of which operations can be merged passed other
operations.  Instead have a simple rule of only merging an operation
with the previous operation affecting the same partition, skipping past
operations affecting other partitions.  This is simple to understand and
simple to implement.  It is also compatible with how operations are
visually applied by replacing the visual partition with the after copy
from the operation.

Merging of operations past other operations on the same partition is
only an issue for the merging of check, label file system, name
partition and new UUID operations because multiple operations can be
queued for the same partition and can be merged past a paste into
operation on the same partition (MERGE_LAST_WITH_ANY on STAT_REAL
partitions).  It is not an issue for resize/move and format operations
on existing partitions because they don't merge operations past other
operations (MERGE_LAST_WITH_PREV on STAT_REAL partitions).  And for
resize/move, delete and format operations on new not yet created
partitions all operations are merged to a single create new operation
(MERGE_LAST_WITH_ANY on STAT_NEW partitions) so there are never multiple
queued operations to incorrectly merge past.  Additionally paste
operations can't be composed on a new not yet created partition.

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
             the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op