Remove redundant status = STAT_NEW in activate_resize() (#117)(#780034)
As activate_resize() constructs and updates the resized_ptn object from
the selected_partition_ptr, via the dialog, it keeps status member the
same.  Therefore if selected_partition_ptr->status == STAT_NEW then it
is also true that resized_ptn->status == STAT_NEW.  Therefore the
assignment is redundant so remove it.  (The removed line was called when
resizing/moving a new primary or logical partition, not yet created on
the drive).

Historical analysis:
The assignment was added by commit "Replace open coded merge of
resize/move into create operation (#755214)" [1].

Before commit [1] activate_resize() open coded forward merging of a
previous create operation with the currently being composed resize/move
on this pending create partition.  It called
temp_partition.Set_Unallocated() to compose an empty space partition
object to be used as the before copy of the partition in the newly
merged create operation.  Set_Unallocated() explicitly sets
status = STAT_REAL.

With commit [1] activate_resize() added a resize/move operation and
elsewhere in the code the operation was backward merged with the earlier
create operation.  This new code constructed and populated part_temp
from the selected_partition_ptr, via the dialog, much like the current
code does.  Set_Unallocated() was no longer called.  Therefore the
partition status value was preserved so the assignment of
status = STAT_NEW was never needed.

[1] 5793c5ac03bea5b4818309e5755803b00feeb016
    Replace open coded merge of resize/move into create operation (#755214)

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
             the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op