Fix crash from assert failure in ped_partition_is_flag_available() (#117)(#780034)
Compose these operations:
(1) Copy existing partition pasting into new,
(2) Format new partition,
and then on the new partition attempt Partition > Manage Flags.  GParted
reports this dialog:

                          Libparted Bug
    Assertion (ped_partition_is_active (part)) at disk.c:1416
    in function ped_partition_is_flag_available() failed.
                                                     [  No  ]

and core dumps like this:
    # ./gpartedbin
    GParted 1.7.0-git
    configuration --enable-libparted-dmraid
    libparted 3.2
    Backtrace has 20 calls on the stack:
    ...
    Assertion (ped_partition_is_active (part)) at disk.c:1416 in function ped_partition_is_flag_available() failed.
    # gdb ./gpartedbin core --batch --quiet --ex backtrace -ex quit
    ...
    Core was generated by `./gpartedbin'.
    Program terminated with signal SIGABRT, Aborted.
    #0  0x00007f8d5e82152f in raise () from /lib64/libc.so.6
    [Current thread is 1 (Thread 0x7f8d63f9d680 (LWP 581834))]
    #0  0x00007f8d5e82152f in raise () at /lib64/libc.so.6
    #1  0x00007f8d5e7f4e65 in abort () at /lib64/libc.so.6
    #2  0x00007f8d5fb13fbc in ped_assert (cond_text=cond_text@entry=0x7f8d5fb3ed50 "ped_partition_is_active (part)", file=file@entry=0x7f8d5fb3e858 "disk.c", line=line@entry=1416, function=function@entry=0x7f8d5fb3f640 <__PRETTY_FUNCTION__.5323> "ped_partition_is_flag_available") at debug.c:111
    #3  0x00007f8d5fb177fa in ped_partition_is_flag_available (part=0x26d0610, flag=PED_PARTITION_BOOT) at disk.c:1416
    #4  0x000000000046b9d1 in GParted::GParted_Core::get_available_flags(GParted::Partition const&) (this=this@entry=0x7fffc8e66990, partition=...) at /usr/include/c++/8/bits/stl_vector.h:930
    #5  0x00000000004a9ba4 in GParted::Win_GParted::activate_manage_flags() (this=0x7fffc8e65f60) at Win_GParted.cc:3277
    ...

GParted is asking libparted what partition flags are available on a
partition which it has not yet created, hence the assert failure.
Bisected this crash to being introduced by commit "Remove superfluous
STAT_FORMATTED" [1] new in GParted 1.6.0.

Prior to commit [1] the status of the partition object in a composed
format operation was set to STAT_NEW when formatting a new not yet
created partition and STAT_FORMATTED otherwise.  This prevented most
operations on top of it, including Manage Flags.  See
set_valid_operations().

After commit [1] the status of the partition was set to STAT_NEW when
formatting a new not yet created partition and STAT_REAL otherwise.
Specifically when formatting a pasted into new partition (STAT_COPY) the
status was changed to STAT_REAL by the temp_filesystem_ptn.Reset() call,
thus allowing Manage Flags on a not yet created partition.

Fix this by making resize/move maintain the partition status the same
as the selected partition (STAT_REAL, _NEW or _COPY) thus preventing
attempting Manage Flags on not yet created partitions.

[1] eb6a77c4fdcd4ae93eec19a9555da57e0879b3b5
    Remove superfluous STAT_FORMATTED

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
             the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op