Maintain partition status while composing format operation (#117)(#780034)
In the case of formatting a closed LUKS encrypted partition, temp_ptn is
default constructed (Partition::Partition()), which calls Reset() which
sets the status to STAT_REAL.  Maintain the partition status by copying
it from the selected partition.  So far this makes no run-time
difference because only existing (STAT_REAL) partitions containing
closed LUKS encrypted data can be formatted.  Create new partition
(STAT_NEW) and copy paste into new partition (STAT_COPY) of closed LUKS
encrypted data is not supported so formatting of these cases can't
arise.  This is being done for consistency so that activate_format()
always maintains the status of the new partition object that is
composes.

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
             the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op