Remove redundant operation merging when deleting a new partition (#117)(#780034)
When a new not yet created partition (STAT_NEW) is deleted GParted
removes all operations for that partition.  Afterwards all newly
adjacent operations are re-tested for mergeability (MERGE_ALL_ADJACENT).
With all other operations now using the new merge rule
(MERGE_LAST_WITH_PREV_SAME_PTN) all mergeable operations have already
been merged and deleting all operations for a not yet created partition
doesn't allow for any more merging of operations on other partitions.

Test case: A disk with a partition and some empty space:
    [  /dev/sdb1  ][  unallocated  ]
Compose these operations:
1. Format /dev/sdb1 as ext2
2. Create primary partition #1 (linux-swap) on /dev/sdb
3. Format /dev/sdb1 as ext3
4. Delete New Partition #1 (linux-swap) from /dev/sdb

Before this patchset GParted queues the first three operations and on
composing the fourth, operations 2 and 4 are removed and 1 and 3 become
adjacent and are merged.

Before this commit GParted queues the first two operations and on
composing the third, it is merged past the second into the first.  Then
on composing the forth, operations 2 and 4 are removed.

Operations are already merged past those for other partitions so
deleting all operations for another partition now doesn't allow for more
operations to be merged.  Therefore remove the redundant merge request.

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
             the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op