Refactor add_operation() into if fail return early pattern (#117)(#780034)
Refactor into if fail return early pattern to make the code easier to
understand and re-write the operation and new partition validity check
to be clearer in which cases it applies.  Also rename the method to the
predominant all lower case.

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
             the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op