Attempt merging of all operations when added (#177)(#780043)
Operation merging is enabled in two different ways.  Whether
Win_GParted::merge_operations() is called when composing each operation
and by what each Operation*::merge_operations() allows.  The later is
the ultimate control because those methods have to implement merging of
the operation or not.  So just always attempt to merge the latest
operation and rely on what each Operation*::merge_operations()
implements.

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
             the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op