Simplify operation merging management (#117)(#780034)
Change from always adding an operation to the list and then attempting
to merge it afterwards to attempt to merge the operation first and if
that is not possible add it instead.  This makes the code simpler by
allowing for the removal of merge_two_operations(), but add_operation()
does have to take on the responsibility for deleting the operation when
it's merged into the list it's not wanted any more.

Bug 780034 - Partlabel/label is not updated if a newer job is queued for
              the same partition
Closes #117 - Partition name is not updated in GUI if another "set
              partition name" op is queued after a non-rename op