Rename member to Win_GParted::m_current_device and list initialise
This is continuing the work towards having all member variables starting
with m_ to make it abundantly clear that they are members of a class.
Also initialise m_current_device in the constructor initialisation list
rather than by assignment in the constructor body [1][2].  It's
recommended because it's more efficient and const and reference members
can only be initialised in the initialisation list.

Just going to rename a few of the Win_GParted member variables that I've
been looking at recently in the next few commits.

[1] C++ FAQ / Should my constructors use "initialization lists" or
    "assignment"?
    https://isocpp.org/wiki/faq/ctors#init-lists
        Initialization lists.  In fact, constructors should initialize
        as a rule all member objects in the initialization list.

[2] C++: Where to initialize variables in constructor [duplicate]
    https://stackoverflow.com/questions/6822422/c-where-to-initialize-variables-in-constructor/6822479