Rename member to Win_GParted::m_operationslist_open and list initialise
GParted uses all upper case for macro names and enumerator identifiers,
not member variables.  Avoid confusion and rename to "snake_case" [1]
and prefix with m_ as it's a member variable.  Also initialise in the
constructor initialisation list as is recommended.

(There are multiple other transgressors to be updated at another time).

[1] Naming convention (programming),
    Examples of multiple-word identifier formats
    https://en.wikipedia.org/wiki/Naming_convention_(programming)#Examples_of_multiple-word_identifier_formats