Replace C strings with C++ std::strings in detect_filesystem_internal()
This simplifies the code and makes it more robust because C++
std::strings are arrays of characters of known length, where as C
strings are NUL byte terminated sequences of characters of unknown
length.
1. No longer have to assume char[16] will be big enough to store the
   largest possible magic values.
2. No longer have to separately track the length of the current
   signatures to be match with variables len1 and len2.
3. Allows for 0 (NUL) character in std::string so can correctly match 2
   byte "\x01\x00" sig1 for APFS.

Also the signatures variable never changes and is used read-only.
Therefore make it static const so it is only initialised once during the
life of the program and read-only access is enforced.