Rename member variable to Dialog_Progress::m_operations
As well as the usual reason of making it abundantly clear it's a member
variable it also avoids confusion reading the constructor code because
it has a parameter with the same name as the member variable.  Which
was meant when 'operations' was used?  (The parameter named operations
shadowed member variable of the same name, hence previous use of
this->operations).  And the initialiser list contained this:
    operations(operations)

The same name used twice, each referring to a different thing.  Giving
the member variable a different name to the constructor parameter avoids
this confusion.