Introduce typedef OperationVector (!145)
It's a short hand for:
    std::vector<std::unique_ptr<Operation>>

As discussed in the previous commit message, modern C++ guidance is that
passing a unique_ptr<T> is transferring ownership.  However passing an
object that internally contains a unique_ptr is not transferring
ownership.  In this case passing a vector of unique_ptr to Operation by
reference is not transferring ownership of the Operation objects, they
are still owned by the vector.  Replacing the long form type with the
shorter typedef name which doesn't contain "unique_ptr" is meant to help
understanding this and be simpler to read and comprehend.

Closes !145 - Convert Operations to smart pointers