Rename member to Operation::m_type and list initialise
Continue the work towards having all member variables start with m_ to
make it very clear they are members of a class.  As recommended use list
initialisation in the constructor rather than assignment in the
constructor body [1].  Base class member variables can not directly be
included in the initialisation list of a derived class constructor, even
if they are public members, because they are not members of the derived
class.  Instead they must be initialised by passing values to a
parameterised base class constructor [2].  Example:
    class Base
    {
    public:
        Base(Value val) : m_val(val);
        Value m_val;
    }

    class Derived : public Base
    {
    public:
        // Derived(Value val) : m_val(val);  // Error: m_val not a member of Derived
        Derived(Value val) : Base(val);
    }

Also move m_type to be the first mentioned member variable in the class
because the type of the operation is the most important characteristic
of an Operation object.

[1] C++ FAQ / Should my constructors use "initialization lists" or
    "assignment"?
    https://isocpp.org/wiki/faq/ctors#init-lists
        Initialization lists.  In fact, constructors should initialize
        as a rule all member objects in the initialization list.

[2] How can I initialize base class member variables in derived class
    constructor?
    https://stackoverflow.com/questions/7405740/how-can-i-initialize-base-class-member-variables-in-derived-class-constructor