Compare using m_partition_new in OperationCheck::merge_operations()
This makes no functional change because the OperationCheck constructor
duplicates the same partition object for m_partition_original and
m_partition_new.  However every other operation and logically speaking
it should be the new partition object of the existing operation which is
compared with the original partition object of the candidate operation
when confirming operation mergeability.  Update accordingly.

Visually:
    Source         Partition object

    from drive ----[sdb1]

    existing    +--[partition_original]
    operation --+
    in list     +--[partition_new]      \
                                        | Compared in merge_operations()
    candidate   +--[partition_original] /
    operation --+
    for merge   +--[partition_new]