Clear ESP when copying non-FAT file system into existing (#288)
Now copy a non-FAT file system over the top of an existing partition
with on-disk type of EFI System Partition (ESP).  GParted previews and
implements this as leaving the ESP flag set.

As non-FAT file system can't be an ESP [1] clear the flag to set a file
system specific partition type when copying a none FAT file system into
an existing partition.

[1] UEFI Specification 2.11, 13.3. File System Format
    https://uefi.org/specs/UEFI/2.11/13_Protocols_Media_Access.html#file-system-format
        EFI encompasses the use of FAT32 for a system partition, and
        FAT12 or FAT16 for removable media.

Closes #288 - GPT partition type GUID (code) does not get copied in
              partition copy & paste feature