Set LVM flag for preview of create LVM2 PV (#288)
When GParted creates a new LVM2 PV partition it sets the libparted LVM
flag to set the on-disk partition type appropriately [1].  However when
previewing the operation the flag is not shown.  Set the LVM flag when
composing create a new LVM2 PV partition so the preview matches what
will be implemented.

Also see earlier commit "Set libparted ESP and LVM flags correctly when
formatting (#288)" for the discussion about choosing to always set the
LVM flag regardless of whether libparted supports it on the partition
table or not.

[1] c3ab62591b73266f43b379d9a7aef3be13f3c384
    Add creation of LVM2 PVs (#670171)

Closes #288 - GPT partition type GUID (code) does not get copied in
              partition copy & paste feature