Set ESP and LVM flags correctly when formatting (#288)
Formatting a partition for all file system types previews as clearing
all flags.  When applied the ESP flag is always cleared regardless of
the file system type.  When applied formatting to LVM2 PV, the LVM flag
is always set.

Fix format composition to do the right thing:
1. Formatting to a FAT file system in an EFI System Partition (ESP)
   keeps the ESP flag.
2. Formatting any other file system in an ESP clears the ESP flag.
   (The code implicitly already does this by clearing all flags).
3. Formatting to an LVM2 PV sets the LVM flag.
4. Formatting to a non-LVM2 PV clears the LVM flag.
   (The code implicitly already does this by clearing all flags).

Note that in libparted the LVM flag is only available for these
partition tables: AMIGA, GPT, MAC, MSDOS & SUN.  It's not available in
the remaining partition tables: ATARI, AIX, BSD, DVH, & PC98.  As the
LVM flag is available for MSDOS and GPT, virtually the only tables used
nowadays, then always set the flag when composing and therefore
previewing the operation.  Don't bother trying to not set the LVM flag
when formatting a partition on one of the other tables, it's not worth
the code complexity.  When the operation is applied the LVM flag will
only be set set if it's supported.

Closes #288 - GPT partition type GUID (code) does not get copied in
              partition copy & paste feature