Clear LVM flag when copying non-LVM2 PV file system into existing (#288)
Copy a non-LVM2 PV file system into an existing partition with the LVM
flag set.  GParted previews this as leaving the LVM flag set but when
applied the LVM flag is cleared.  Fix this so the preview of the copy
operation clears the LVM flag.  Copying an LVM2 PV is not supported so
just always clear the LVM flag when pasting into an existing partition.

Closes #288 - GPT partition type GUID (code) does not get copied in
              partition copy & paste feature