Fix expansion of details in Applying pending operations dialog (#168)
The operation details widget didn't resize vertically when the Applying
pending operations dialog was resized.  Instead it remained a height
size and extra space in the dialog remained unused.  That has been the
case since GParted 1.0.0 and the Gtk3 port.

Fix this by setting vexpand on the operation details so the widget
expands vertically when the dialog is resized.

[1] gtkmm - How to make box containing widgets expand to fill window?
    https://stackoverflow.com/questions/49088760/gtkmm-how-to-make-box-containing-widgets-expand-to-fill-window

Closes #168 - Progress window can be resized, but Details height doesn't
              change