Generate a warning for logical EFI System Partitions (ESP)s (#291)
A user was converting their laptop to boot in UEFI mode.  On a Master
Boot Record (MBR) partitioned drive they created an EFI System Partition
(ESP) in a logical partition but their laptop didn't boot.  When the
ESP was re-create in a primary partition their laptop did boot.  They
requested that GParted warn when a logical partition is set as an ESP.

The UEFI Specification [1] describes the MBR partition table that is
supported but there is no mention of primary, extended or logical
partitions.  Only partition type 0xEF (EFI System Partition) in the MBR
in sector 0.

After also finding these results [2][3], the conclusion is that UEFI
booting from a logical ESP is not supported.  Therefore make GParted
generate a partition warning when a logical partition is set as an ESP.

[1] UEFI Specification 2.11, 5.2.1 Legacy Master Boot Record (MBR)
    https://uefi.org/specs/UEFI/2.11/05_GUID_Partition_Table_Format.html#legacy-master-boot-record-mbr
        A legacy MBR may be located at LBA 0 (i.e., the first logical
        block) of the disk if it is not using the GPT disk layout (i.e.,
        if it is using the MBR disk layout).  The boot code on the MBR
        is not executed by UEFI firmware.
        ...
        The MBR contains four partition records (see Table 11) that each
        define the beginning and ending LBAs that a partition consumes
        on a disk.

[2] Can the EFI System Partition be a logical partition on MBR disks?
    https://unix.stackexchange.com/questions/277858/can-the-efi-system-partition-be-a-logical-partition-on-mbr-disks
        The specification in section 5, starting on PDF page 165, gives
        the layout of an MBR partition table.  It does not mention the
        extended partition, so UEFI does not support it.

        Thus, according to the standard, would appear to be "no".

[3] Should ESP (EFI) partition be primary or logical?
    https://superuser.com/questions/1410368/should-esp-efi-partition-be-primary-or-logical
        I want to create an ESP partition for my EFI on a new disk.

        Should I check "Create as Primary Partition"?

        I found out: yes, I should.  Otherwise the partition may risk
        not being seen at startup.

Closes #291 - Warn about EFI System set on a logical partition